feat: adding the request context
This feature handle the SQL queries that are taking longer to run than
expected. When this happens, it suggests a problem — either with that
particular query or our database or application more generally — and we
probably want to cancel the query (in order to free up resources), log
an error for further investigation, and return a 500 Internal Server
Error response to the client.